Refactor info processing to its own class
